64 bits compile fix
svn:sync-xref-src-repo-rev: r20290

+++ src/VBox/VMM/VMMAll/EMAll.cpp
@@ -1,4 +1,4 @@
-/* $Id: EMAll.cpp 2000 2007-04-10 08:02:14Z noreply@oracle.com $ */
+/* $Id: EMAll.cpp 2002 2007-04-10 09:00:21Z noreply@oracle.com $ */
 /** @file
  * EM - Execution Monitor(/Manager) - All contexts
  */
@@ -50,7 +50,7 @@
 *   Structures and Typedefs                                                    *
 *******************************************************************************/
 typedef EMDECL(uint32_t) PFN_EMULATE_PARAM2(uint32_t *pu32Param1, uint32_t val2);
-typedef EMDECL(uint32_t) PFN_EMULATE_PARAM3(uint32_t *pu32Param1, uint32_t val2, uint32_t val3);
+typedef EMDECL(uint32_t) PFN_EMULATE_PARAM3(uint32_t *pu32Param1, uint32_t val2, size_t val3);
 
 /*******************************************************************************
  *   Internal Functions                                                        *
